requirements.txt
pymavlink
pyzmq
msgpack
pyserial